+++ src/integration.ts
@@ -1,6 +1,5 @@
 import ts from "typescript";
 import { z } from "zod";
-import { ZodFile } from "./file-schema";
 import {
   emptyHeading,
   emptyTail,
@@ -27,7 +26,7 @@ import {
   protectedReadonlyModifier,
   spacingMiddle,
 } from "./integration-helpers";
-import { defaultSerializer, hasRaw, makeCleanId } from "./common-helpers";
+import { defaultSerializer, makeCleanId } from "./common-helpers";
 import { Method, methods } from "./method";
 import { mimeJson } from "./mime";
 import { loadPeer } from "./peer-helpers";
@@ -155,11 +154,10 @@ export class Integration {
           makeAlias: this.makeAlias.bind(this),
           optionalPropStyle,
         };
-        const inputSchema = endpoint.getSchema("input");
         const inputId = makeCleanId(method, path, "input");
         const input = zodToTs({
           ...commons,
-          schema: hasRaw(inputSchema) ? ZodFile.create().buffer() : inputSchema,
+          schema: endpoint.getSchema("input"),
           isResponse: false,
         });
         const positiveResponseId = splitResponse