+++ tests/unit/file-schema.spec.ts
@@ -1,90 +1,78 @@
-import { ZodFile } from "../../src/file-schema";
+import { expectType } from "tsd";
+import { z } from "zod";
+import { getMeta } from "../../src/metadata";
+import { ez } from "../../src";
 import { readFile } from "node:fs/promises";
 import { describe, expect, test } from "vitest";
 
-describe("ZodFile", () => {
-  describe("static::create()", () => {
+describe("ez.file()", () => {
+  describe("creation", () => {
     test("should create an instance being string by default", () => {
-      const schema = ZodFile.create();
-      expect(schema).toBeInstanceOf(ZodFile);
-      expect(schema._def.encoding).toBeUndefined();
-      expect(schema._def.typeName).toEqual("ZodFile");
-      expect(schema.isBinary).toBeFalsy();
-      expect(schema.isBase64).toBeFalsy();
-      expect(schema.isString).toBeTruthy();
-      expect(schema.isBuffer).toBeFalsy();
+      const schema = ez.file();
+      expect(schema).toBeInstanceOf(z.ZodString);
+      expect(getMeta(schema, "kind")).toBe("File");
     });
+
+    test.each([ez.file("string"), ez.file().string("deprecated message")])(
+      "should create a string file",
+      (schema) => {
+        expect(schema).toBeInstanceOf(z.ZodString);
+        expectType<string>(schema._output);
+      },
+    );
+
+    test.each([ez.file("buffer"), ez.file().buffer("deprecated message")])(
+      "should create a buffer file",
+      (schema) => {
+        expect(schema).toBeInstanceOf(z.ZodEffects);
+        expectType<Buffer>(schema._output);
+      },
+    );
+
+    test.each([ez.file("binary"), ez.file().binary("deprecated message")])(
+      "should create a binary file",
+      (schema) => {
+        expect(schema).toBeInstanceOf(z.ZodUnion);
+        expectType<Buffer | string>(schema._output);
+      },
+    );
+
+    test.each([ez.file("base64"), ez.file().base64("deprecated message")])(
+      "should create a base64 file",
+      (schema) => {
+        expect(schema).toBeInstanceOf(z.ZodString);
+        expectType<string>(schema._output);
+      },
+    );
   });
 
-  describe(".string()", () => {
-    test("should create a string file", () => {
-      const schema = ZodFile.create().string();
-      expect(schema).toBeInstanceOf(ZodFile);
-      expect(schema._def.encoding).toBeUndefined();
-      expect(schema.isString).toBeTruthy();
-      expect(schema.isBuffer).toBeFalsy();
-    });
-  });
-
-  describe(".buffer()", () => {
-    test("should create a buffer file", () => {
-      const schema = ZodFile.create().buffer();
-      expect(schema).toBeInstanceOf(ZodFile);
-      expect(schema._def.encoding).toBeUndefined();
-      expect(schema.isBuffer).toBeTruthy();
-      expect(schema.isString).toBeFalsy();
-    });
-  });
-
-  describe(".binary()", () => {
-    test("should create a binary file", () => {
-      const schema = ZodFile.create().binary("test message");
-      expect(schema).toBeInstanceOf(ZodFile);
-      expect(schema.isBinary).toBeTruthy();
-      expect(schema._def.encoding).toBe("binary");
-      expect(schema._def.message).toBe("test message");
-    });
-  });
-
-  describe(".base64()", () => {
-    test("should create a base64 file", () => {
-      const schema = ZodFile.create().base64("test message");
-      expect(schema).toBeInstanceOf(ZodFile);
-      expect(schema.isBase64).toBeTruthy();
-      expect(schema._def.encoding).toBe("base64");
-      expect(schema._def.message).toBe("test message");
-    });
-  });
-
-  describe("_parse()", () => {
+  describe("parsing", () => {
     test.each([
       {
-        schema: ZodFile.create(),
+        schema: ez.file(),
         subject: 123,
+        code: "invalid_type",
         expected: "string",
         received: "number",
         message: "Expected string, received number",
       },
       {
-        schema: ZodFile.create().buffer(),
+        schema: ez.file("buffer"),
         subject: "123",
-        expected: "object",
-        received: "string",
+        code: "custom",
         message: "Expected Buffer",
+        fatal: true,
       },
     ])(
       "should invalidate wrong types",
-      ({ schema, subject, expected, received, message }) => {
+      ({ schema, subject, ...expectedError }) => {
         const result = schema.safeParse(subject);
         expect(result.success).toBeFalsy();
         if (!result.success) {
           expect(result.error.issues).toEqual([
             {
-              code: "invalid_type",
-              expected,
-              message,
+              ...expectedError,
               path: [],
-              received,
             },
           ]);
         }
@@ -92,14 +80,15 @@ describe("ZodFile", () => {
     );
 
     test("should perform additional check for base64 file", () => {
-      const schema = ZodFile.create().base64();
+      const schema = ez.file("base64");
       const result = schema.safeParse("~~~~");
       expect(result.success).toBeFalsy();
       if (!result.success) {
         expect(result.error.issues).toEqual([
           {
-            code: "custom",
+            code: "invalid_string",
             message: "Does not match base64 encoding",
+            validation: "regex",
             path: [],
           },
         ]);
@@ -107,7 +96,7 @@ describe("ZodFile", () => {
     });
 
     test("should accept string", () => {
-      const schema = ZodFile.create();
+      const schema = ez.file();
       const result = schema.safeParse("some string");
       expect(result).toEqual({
         success: true,
@@ -116,7 +105,7 @@ describe("ZodFile", () => {
     });
 
     test("should accept Buffer", () => {
-      const schema = ZodFile.create().buffer();
+      const schema = ez.file("buffer");
       const subject = Buffer.from("test", "utf-8");
       const result = schema.safeParse(subject);
       expect(result).toEqual({
@@ -126,7 +115,7 @@ describe("ZodFile", () => {
     });
 
     test("should accept binary read string", async () => {
-      const schema = ZodFile.create().binary();
+      const schema = ez.file("binary");
       const data = await readFile("logo.svg", "binary");
       const result = schema.safeParse(data);
       expect(result).toEqual({
@@ -136,7 +125,7 @@ describe("ZodFile", () => {
     });
 
     test("should accept base64 read string", async () => {
-      const schema = ZodFile.create().base64();
+      const schema = ez.file("base64");
       const data = await readFile("logo.svg", "base64");
       const result = schema.safeParse(data);
       expect(result).toEqual({